refactored DeviceService

+++ src/main/java/org/buyandfly/service/IService.java
@@ -3,7 +3,7 @@ package org.buyandfly.service;
 import java.util.List;
 
 public interface IService<T> {
-    T getById(final int id);
+    T getById(final long id);
     List<T> getAll();
     void add(final T t);
     void edit(final T t);
